streamlit/streamlit_requirements.txt
streamlit==1.28.0
pandas==2.0.3
requests==2.31.0
plotly==5.17.0
biopython==1.81
numpy==1.24.3